infra/images/musetalk-worker/render.py
"""MuseTalk render engine — lip-sync avatar video generation.

Loads MuseTalk model in FP16 for V100 compatibility.
Input: reference image + audio file path
Output: rendered MP4 path
"""

import os
import hashlib
import logging
import time
import torch

log = logging.getLogger("render")

_model = None
_device = None


def get_cache_key(persona_id: str, audio_hash: str) -> str:
    """Deterministic cache key for rendered clips."""
    return hashlib.sha256(f"{persona_id}:{audio_hash}".encode()).hexdigest()[:16]


def load_model():
    """Lazy-load MuseTalk model to GPU with FP16."""
    global _model, _device

    if _model is not None:
        return _model

    cache_dir = os.environ.get("MODEL_CACHE_DIR", "/models")
    dtype_str = os.environ.get("MODEL_DTYPE", "float16")
    dtype = torch.float16 if dtype_str == "float16" else torch.float32

    _device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
    log.info("Loading MuseTalk on %s with dtype=%s", _device, dtype)

    # Import MuseTalk modules from downloaded repo
    import sys
    sys.path.insert(0, os.path.join(cache_dir, "musetalk"))

    from musetalk.utils.utils import load_all_model
    from musetalk.utils.preprocessing import get_landmark_and_bbox

    _model = {
        "models": load_all_model(),
        "get_landmark": get_landmark_and_bbox,
        "dtype": dtype,
        "device": _device,
    }

    log.info("MuseTalk loaded successfully (VRAM: %.1f MB used)",
             torch.cuda.memory_allocated() / 1024 / 1024)
    return _model


def render_avatar(
    reference_image_path: str,
    audio_path: str,
    output_path: str,
    fps: int = 25,
) -> dict:
    """Render a lip-synced avatar video.

    Returns dict with timing and output metadata.
    """
    t0 = time.time()
    model = load_model()

    # MuseTalk inference pipeline:
    # 1. Extract face landmarks from reference image
    # 2. Process audio into mel spectrogram chunks
    # 3. For each chunk, generate mouth region via latent inpainting
    # 4. Composite onto reference frame
    # 5. Encode to MP4

    from musetalk.utils.utils import datagen
    from musetalk.utils.preprocessing import get_landmark_and_bbox, coord_placeholder

    audio_feats, _ = model["models"]["audio_processor"].audio2feat(audio_path)
    # ... (full MuseTalk pipeline will be integrated here)
    # For now this is a skeleton that will be filled with the actual
    # MuseTalk inference loop once we validate the container builds

    elapsed = time.time() - t0
    return {
        "output_path": output_path,
        "render_time_s": round(elapsed, 2),
        "gpu": str(model["device"]),
        "dtype": str(model["dtype"]),
    }
