infra/images/musetalk-worker/requirements.txt
# MuseTalk worker dependencies
# Pinned for CUDA 11.8 + V100 compatibility (FP16, no BF16)

# MuseTalk core
diffusers==0.28.2
transformers==4.42.4
accelerate==0.31.0
safetensors==0.4.3
opencv-python-headless==4.10.0.84
Pillow==10.4.0
scipy==1.13.1
numpy==1.26.4

# Audio processing
librosa==0.10.2
soundfile==0.12.1

# Face detection / alignment (mmlab stack)
mmcv==2.1.0
mmdet==3.3.0
mmpose==1.3.1
mmengine==0.10.4

# Video processing
moviepy==1.0.3
imageio==2.34.2
imageio-ffmpeg==0.5.1

# NATS client (JetStream support)
nats-py==2.7.2

# HTTP server for health checks
aiohttp==3.9.5

# Prometheus metrics
prometheus-client==0.20.0

# HuggingFace Hub for model downloads
huggingface-hub==0.23.4

# S3-compatible storage
boto3==1.34.131
